adds comments and updates PinPositive to Pin
